+++ SonLVLAPI/S2/Object.cs
@@ -62,8 +62,8 @@ namespace SonicRetro.SonLVL.API.S2
 			if (YFlip) val |= 0x4000;
 			if (RememberState) val |= 0x8000;
 			ret.AddRange(ByteConverter.GetBytes(val));
-			ret.Add(ID);
-			ret.Add(SubType);
+			ret.Add((byte)ID);
+			ret.Add((byte)SubType);
 			return ret.ToArray();
 		}
